Add first failing feature test for list screencasts
Fist I create a file in the feature directory in the app.

Using a file ending with .feature, define what concept this
feature will be about. This is only for noting the context.
This "Feature:" description can be several lines and it's
common to have it be a short name and a breif narrative.
It is also common to see this bareif narrative after a
emplty line or indented. Cucumber does not care what you
write here. It is only looking for keywords like "Scenario"

After features, scenarios are given (generally the "happy
path" first followed by edge cases). The "Scenario:" is
followed by steps.

There number of steps is arbitrary and begins with the step
keywords "Given", "When", "Then", "And", and "But".

To run this test use:
```rake cucumber``` or ```bundle exec cucumber```

You should see the output of the test which will also show
you suggestions for implementing step definitions.

Run this test