Create step file and first step for list screencast Scenario
There is no definition for "I am a guest" as that is the defult
user.

For a screencast exists we use a new Catalog object and stub out
the :all method that shoudl return an array of hashes.

When this test is ran, we see that Catalog does not exits.
This helps us build out our application.